Moved functions from rpc.rs into the shared crate.
