+++ apps/src/lib/client/rpc.rs
@@ -53,7 +53,7 @@ use tokio::time::{Duration, Instant};
 
 use namada::ledger::wallet::Wallet;
 use crate::cli::{self, args};
-use crate::client::tendermint_rpc_types::TxResponse;
+use namada::ledger::rpc::TxResponse;
 use namada::ledger::masp::{Conversions, PinnedBalanceError};
 use crate::facade::tendermint::merkle::proof::Proof;
 use crate::facade::tendermint_rpc::error::Error as TError;
@@ -68,13 +68,13 @@ use crate::facade::tendermint_rpc::{
 /// error.
 pub async fn query_tx_status<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
-    status: TxEventQuery<'_>,
+    status: namada::ledger::rpc::TxEventQuery<'_>,
     deadline: Instant,
 ) -> Event {
     const ONE_SECOND: Duration = Duration::from_secs(1);
     // sleep for the duration of `backoff`,
     // and update the underlying value
-    async fn sleep_update(query: TxEventQuery<'_>, backoff: &mut Duration) {
+    async fn sleep_update(query: namada::ledger::rpc::TxEventQuery<'_>, backoff: &mut Duration) {
         tracing::debug!(
             ?query,
             duration = ?backoff,
@@ -114,28 +114,19 @@ pub async fn query_tx_status<C: Client + namada::ledger::queries::Client + Sync>
 
 /// Query the epoch of the last committed block
 pub async fn query_epoch<C: Client + namada::ledger::queries::Client + Sync>(client: &C) -> Epoch {
-    let epoch = unwrap_client_response::<C, _>(RPC.shell().epoch(client).await);
-    println!("Last committed epoch: {}", epoch);
-    epoch
+    namada::ledger::rpc::query_epoch(client).await
 }
 
 /// Query the last committed block
 pub async fn query_block<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
 ) -> crate::facade::tendermint_rpc::endpoint::block::Response {
-    let response = client.latest_block().await.unwrap();
-    println!(
-        "Last committed block ID: {}, height: {}, time: {}",
-        response.block_id,
-        response.block.header.height,
-        response.block.header.time
-    );
-    response
+    namada::ledger::rpc::query_block(client).await
 }
 
 /// Query the results of the last committed block
 pub async fn query_results<C: Client + namada::ledger::queries::Client + Sync>(client: &C) -> Vec<BlockResults> {
-    unwrap_client_response::<C, _>(RPC.shell().read_results(client).await)
+    namada::ledger::rpc::query_results(client).await
 }
 
 /// Query the specified accepted transfers from the ledger
@@ -988,8 +979,7 @@ pub async fn get_token_balance<C: Client + namada::ledger::queries::Client + Syn
     token: &Address,
     owner: &Address,
 ) -> Option<token::Amount> {
-    let balance_key = balance_key(token, owner);
-    query_storage_value(client, &balance_key).await
+    namada::ledger::rpc::get_token_balance(client, token, owner).await
 }
 
 pub async fn query_proposal_result<C: Client + namada::ledger::queries::Client + Sync>(
@@ -1779,8 +1769,7 @@ pub async fn get_public_key<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
     address: &Address,
 ) -> Option<common::PublicKey> {
-    let key = pk_key(address);
-    query_storage_value(client, &key).await
+    namada::ledger::rpc::get_public_key(client, address).await
 }
 
 /// Check if the given address is a known validator.
@@ -1788,7 +1777,7 @@ pub async fn is_validator<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
     address: &Address,
 ) -> bool {
-    unwrap_client_response::<C, _>(RPC.vp().pos().is_validator(client, address).await)
+    namada::ledger::rpc::is_validator(client, address).await
 }
 
 /// Check if a given address is a known delegator
@@ -1796,10 +1785,7 @@ pub async fn is_delegator<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
     address: &Address,
 ) -> bool {
-    let bonds_prefix = pos::bonds_for_source_prefix(address);
-    let bonds =
-        query_storage_prefix::<C, pos::Bonds>(&client, &bonds_prefix).await;
-    bonds.is_some() && bonds.unwrap().count() > 0
+    namada::ledger::rpc::is_delegator(client, address).await
 }
 
 pub async fn is_delegator_at<C: Client + namada::ledger::queries::Client + Sync>(
@@ -1807,13 +1793,7 @@ pub async fn is_delegator_at<C: Client + namada::ledger::queries::Client + Sync>
     address: &Address,
     epoch: Epoch,
 ) -> bool {
-    let key = pos::bonds_for_source_prefix(address);
-    let bonds_iter = query_storage_prefix::<C, pos::Bonds>(client, &key).await;
-    if let Some(mut bonds) = bonds_iter {
-        bonds.any(|(_, bond)| bond.get(epoch).is_some())
-    } else {
-        false
-    }
+    namada::ledger::rpc::is_delegator_at(client, address, epoch).await
 }
 
 /// Check if the address exists on chain. Established address exists if it has a
@@ -1823,14 +1803,7 @@ pub async fn known_address<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
     address: &Address,
 ) -> bool {
-    match address {
-        Address::Established(_) => {
-            // Established account exists if it has a VP
-            let key = storage::Key::validity_predicate(address);
-            query_has_storage_key(client, &key).await
-        }
-        Address::Implicit(_) | Address::Internal(_) => true,
-    }
+    namada::ledger::rpc::known_address(client, address).await
 }
 
 /// Accumulate slashes starting from `epoch_start` until (optionally)
@@ -2039,9 +2012,7 @@ pub async fn query_conversion<C: Client + namada::ledger::queries::Client + Sync
     masp_primitives::transaction::components::Amount,
     MerklePath<Node>,
 )> {
-    Some(unwrap_client_response::<C, _>(
-        RPC.shell().read_conversion(client, &asset_type).await,
-    ))
+    namada::ledger::rpc::query_conversion(client, asset_type).await
 }
 
 /// Query a storage value and decode it with [`BorshDeserialize`].
@@ -2052,34 +2023,7 @@ pub async fn query_storage_value<C: Client + namada::ledger::queries::Client + S
 where
     T: BorshDeserialize,
 {
-    // In case `T` is a unit (only thing that encodes to 0 bytes), we have to
-    // use `storage_has_key` instead of `storage_value`, because `storage_value`
-    // returns 0 bytes when the key is not found.
-    let maybe_unit = T::try_from_slice(&[]);
-    if let Ok(unit) = maybe_unit {
-        return if unwrap_client_response::<C, _>(
-            RPC.shell().storage_has_key(client, key).await,
-        ) {
-            Some(unit)
-        } else {
-            None
-        };
-    }
-
-    let response = unwrap_client_response::<C, _>(
-        RPC.shell()
-            .storage_value(client, None, None, false, key)
-            .await,
-    );
-    if response.data.is_empty() {
-        return None;
-    }
-    T::try_from_slice(&response.data[..])
-        .map(Some)
-        .unwrap_or_else(|err| {
-            eprintln!("Error decoding the value: {}", err);
-            cli::safe_exit(1)
-        })
+    namada::ledger::rpc::query_storage_value(client, key).await
 }
 
 /// Query a storage value and the proof without decoding.
@@ -2089,17 +2033,7 @@ pub async fn query_storage_value_bytes<C: Client + namada::ledger::queries::Clie
     height: Option<BlockHeight>,
     prove: bool,
 ) -> (Option<Vec<u8>>, Option<Proof>) {
-    let data = None;
-    let response = unwrap_client_response::<C, _>(
-        RPC.shell()
-            .storage_value(client, data, height, prove, key)
-            .await,
-    );
-    if response.data.is_empty() {
-        (None, response.proof)
-    } else {
-        (Some(response.data), response.proof)
-    }
+    namada::ledger::rpc::query_storage_value_bytes(client, key, height, prove).await
 }
 
 /// Query a range of storage values with a matching prefix and decode them with
@@ -2112,29 +2046,7 @@ pub async fn query_storage_prefix<C: Client + namada::ledger::queries::Client + 
 where
     T: BorshDeserialize,
 {
-    let values = unwrap_client_response::<C, _>(
-        RPC.shell()
-            .storage_prefix(client, None, None, false, key)
-            .await,
-    );
-    let decode =
-        |PrefixValue { key, value }: PrefixValue| match T::try_from_slice(
-            &value[..],
-        ) {
-            Err(err) => {
-                eprintln!(
-                    "Skipping a value for key {}. Error in decoding: {}",
-                    key, err
-                );
-                None
-            }
-            Ok(value) => Some((key, value)),
-        };
-    if values.data.is_empty() {
-        None
-    } else {
-        Some(values.data.into_iter().filter_map(decode))
-    }
+    namada::ledger::rpc::query_storage_prefix(client, key).await
 }
 
 /// Query to check if the given storage key exists.
@@ -2142,151 +2054,37 @@ pub async fn query_has_storage_key<C: Client + namada::ledger::queries::Client +
     client: &C,
     key: &storage::Key,
 ) -> bool {
-    unwrap_client_response::<C, _>(RPC.shell().storage_has_key(client, key).await)
-}
-
-/// Represents a query for an event pertaining to the specified transaction
-#[derive(Debug, Copy, Clone)]
-pub enum TxEventQuery<'a> {
-    Accepted(&'a str),
-    Applied(&'a str),
-}
-
-impl<'a> TxEventQuery<'a> {
-    /// The event type to which this event query pertains
-    fn event_type(self) -> &'static str {
-        match self {
-            TxEventQuery::Accepted(_) => "accepted",
-            TxEventQuery::Applied(_) => "applied",
-        }
-    }
-
-    /// The transaction to which this event query pertains
-    fn tx_hash(self) -> &'a str {
-        match self {
-            TxEventQuery::Accepted(tx_hash) => tx_hash,
-            TxEventQuery::Applied(tx_hash) => tx_hash,
-        }
-    }
-}
-
-/// Transaction event queries are semantically a subset of general queries
-impl<'a> From<TxEventQuery<'a>> for Query {
-    fn from(tx_query: TxEventQuery<'a>) -> Self {
-        match tx_query {
-            TxEventQuery::Accepted(tx_hash) => {
-                Query::default().and_eq("accepted.hash", tx_hash)
-            }
-            TxEventQuery::Applied(tx_hash) => {
-                Query::default().and_eq("applied.hash", tx_hash)
-            }
-        }
-    }
+    namada::ledger::rpc::query_has_storage_key(client, key).await
 }
 
 /// Call the corresponding `tx_event_query` RPC method, to fetch
 /// the current status of a transation.
 pub async fn query_tx_events<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
-    tx_event_query: TxEventQuery<'_>,
+    tx_event_query: namada::ledger::rpc::TxEventQuery<'_>,
 ) -> std::result::Result<Option<Event>, <C as namada::ledger::queries::Client>::Error> {
-    let tx_hash: Hash = tx_event_query.tx_hash().try_into().unwrap();
-    match tx_event_query {
-        TxEventQuery::Accepted(_) => RPC
-            .shell()
-            .accepted(client, &tx_hash)
-            .await
-            /*.wrap_err_with(|| {
-                eyre!("Failed querying whether a transaction was accepted")
-            })*/,
-        TxEventQuery::Applied(_) => RPC
-            .shell()
-            .applied(client, &tx_hash)
-            .await
-            /*.wrap_err_with(|| {
-                eyre!("Error querying whether a transaction was applied")
-            })*/,
-    }
+    namada::ledger::rpc::query_tx_events(client, tx_event_query).await
 }
 
 /// Lookup the full response accompanying the specified transaction event
 // TODO: maybe remove this in favor of `query_tx_status`
-pub async fn query_tx_response(
-    client: &WebSocketClient,
-    tx_query: TxEventQuery<'_>,
+pub async fn query_tx_response<C: Client + Sync>(
+    client: &C,
+    tx_query: namada::ledger::rpc::TxEventQuery<'_>,
 ) -> Result<TxResponse, TError> {
-    // Find all blocks that apply a transaction with the specified hash
-    let blocks = &client
-        .block_search(tx_query.into(), 1, 255, Order::Ascending)
-        .await
-        .expect("Unable to query for transaction with given hash")
-        .blocks;
-    // Get the block results corresponding to a block to which
-    // the specified transaction belongs
-    let block = &blocks
-        .get(0)
-        .ok_or_else(|| {
-            TError::server(
-                "Unable to find a block applying the given transaction"
-                    .to_string(),
-            )
-        })?
-        .block;
-    let response_block_results = client
-        .block_results(block.header.height)
-        .await
-        .expect("Unable to retrieve block containing transaction");
-    // Search for the event where the specified transaction is
-    // applied to the blockchain
-    let query_event_opt =
-        response_block_results.end_block_events.and_then(|events| {
-            events
-                .iter()
-                .find(|event| {
-                    event.type_str == tx_query.event_type()
-                        && event.attributes.iter().any(|tag| {
-                            tag.key.as_ref() == "hash"
-                                && tag.value.as_ref() == tx_query.tx_hash()
-                        })
-                })
-                .cloned()
-        });
-    let query_event = query_event_opt.ok_or_else(|| {
-        TError::server(
-            "Unable to find the event corresponding to the specified \
-             transaction"
-                .to_string(),
-        )
-    })?;
-    // Reformat the event attributes so as to ease value extraction
-    let event_map: std::collections::HashMap<&str, &str> = query_event
-        .attributes
-        .iter()
-        .map(|tag| (tag.key.as_ref(), tag.value.as_ref()))
-        .collect();
-    // Summarize the transaction results that we were searching for
-    let result = TxResponse {
-        info: event_map["info"].to_string(),
-        log: event_map["log"].to_string(),
-        height: event_map["height"].to_string(),
-        hash: event_map["hash"].to_string(),
-        code: event_map["code"].to_string(),
-        gas_used: event_map["gas_used"].to_string(),
-        initialized_accounts: serde_json::from_str(
-            event_map["initialized_accounts"],
-        )
-        .unwrap_or_default(),
-    };
-    Ok(result)
+    namada::ledger::rpc::query_tx_response(client, tx_query).await
 }
 
 /// Lookup the results of applying the specified transaction to the
 /// blockchain.
-pub async fn query_result(client: &WebSocketClient, args: args::QueryResult) {
+pub async fn query_result<C: Client + Sync>(
+    client: &C,
+    args: args::QueryResult,
+) {
     // First try looking up application event pertaining to given hash.
     let tx_response = query_tx_response(
         client,
-        TxEventQuery::Applied(&args.tx_hash),
+        namada::ledger::rpc::TxEventQuery::Applied(&args.tx_hash),
     )
     .await;
     match tx_response {
@@ -2300,7 +2098,7 @@ pub async fn query_result(client: &WebSocketClient, args: args::QueryResult) {
             // If this fails then instead look for an acceptance event.
             let tx_response = query_tx_response(
                 client,
-                TxEventQuery::Accepted(&args.tx_hash),
+                namada::ledger::rpc::TxEventQuery::Accepted(&args.tx_hash),
             )
             .await;
             match tx_response {
@@ -2323,66 +2121,7 @@ pub async fn get_proposal_votes<C: Client + namada::ledger::queries::Client + Sy
     epoch: Epoch,
     proposal_id: u64,
 ) -> Votes {
-    let validators = get_all_validators(client, epoch).await;
-
-    let vote_prefix_key =
-        gov_storage::get_proposal_vote_prefix_key(proposal_id);
-    let vote_iter =
-        query_storage_prefix::<C, ProposalVote>(client, &vote_prefix_key).await;
-
-    let mut yay_validators: HashMap<Address, VotePower> = HashMap::new();
-    let mut yay_delegators: HashMap<Address, HashMap<Address, VotePower>> =
-        HashMap::new();
-    let mut nay_delegators: HashMap<Address, HashMap<Address, VotePower>> =
-        HashMap::new();
-
-    if let Some(vote_iter) = vote_iter {
-        for (key, vote) in vote_iter {
-            let voter_address = gov_storage::get_voter_address(&key)
-                .expect("Vote key should contain the voting address.")
-                .clone();
-            if vote.is_yay() && validators.contains(&voter_address) {
-                let amount: VotePower =
-                    get_validator_stake(client, epoch, &voter_address)
-                        .await
-                        .into();
-                yay_validators.insert(voter_address, amount);
-            } else if !validators.contains(&voter_address) {
-                let validator_address =
-                    gov_storage::get_vote_delegation_address(&key)
-                        .expect(
-                            "Vote key should contain the delegation address.",
-                        )
-                        .clone();
-                let delegator_token_amount = get_bond_amount_at(
-                    client,
-                    &voter_address,
-                    &validator_address,
-                    epoch,
-                )
-                .await;
-                if let Some(amount) = delegator_token_amount {
-                    if vote.is_yay() {
-                        let entry =
-                            yay_delegators.entry(voter_address).or_default();
-                        entry
-                            .insert(validator_address, VotePower::from(amount));
-                    } else {
-                        let entry =
-                            nay_delegators.entry(voter_address).or_default();
-                        entry
-                            .insert(validator_address, VotePower::from(amount));
-                    }
-                }
-            }
-        }
-    }
-
-    Votes {
-        yay_validators,
-        yay_delegators,
-        nay_delegators,
-    }
+    namada::ledger::rpc::get_proposal_votes(client, epoch, proposal_id).await
 }
 
 pub async fn get_proposal_offline_votes<C: Client + namada::ledger::queries::Client + Sync>(
@@ -2520,53 +2259,7 @@ pub async fn compute_tally<C: Client + namada::ledger::queries::Client + Sync>(
     epoch: Epoch,
     votes: Votes,
 ) -> ProposalResult {
-    let total_staked_tokens: VotePower =
-        get_total_staked_tokens(client, epoch).await.into();
-
-    let Votes {
-        yay_validators,
-        yay_delegators,
-        nay_delegators,
-    } = votes;
-
-    let mut total_yay_staked_tokens = VotePower::from(0_u64);
-    for (_, amount) in yay_validators.clone().into_iter() {
-        total_yay_staked_tokens += amount;
-    }
-
-    // YAY: Add delegator amount whose validator didn't vote / voted nay
-    for (_, vote_map) in yay_delegators.iter() {
-        for (validator_address, vote_power) in vote_map.iter() {
-            if !yay_validators.contains_key(validator_address) {
-                total_yay_staked_tokens += vote_power;
-            }
-        }
-    }
-
-    // NAY: Remove delegator amount whose validator validator vote yay
-    for (_, vote_map) in nay_delegators.iter() {
-        for (validator_address, vote_power) in vote_map.iter() {
-            if yay_validators.contains_key(validator_address) {
-                total_yay_staked_tokens -= vote_power;
-            }
-        }
-    }
-
-    if total_yay_staked_tokens >= (total_staked_tokens / 3) * 2 {
-        ProposalResult {
-            result: TallyResult::Passed,
-            total_voting_power: total_staked_tokens,
-            total_yay_power: total_yay_staked_tokens,
-            total_nay_power: 0,
-        }
-    } else {
-        ProposalResult {
-            result: TallyResult::Rejected,
-            total_voting_power: total_staked_tokens,
-            total_yay_power: total_yay_staked_tokens,
-            total_nay_power: 0,
-        }
-    }
+    namada::ledger::rpc::compute_tally(client, epoch, votes).await
 }
 
 pub async fn get_bond_amount_at<C: Client + namada::ledger::queries::Client + Sync>(
@@ -2575,71 +2268,21 @@ pub async fn get_bond_amount_at<C: Client + namada::ledger::queries::Client + Sy
     validator: &Address,
     epoch: Epoch,
 ) -> Option<token::Amount> {
-    let slashes_key = pos::validator_slashes_key(validator);
-    let slashes = query_storage_value::<C, pos::Slashes>(client, &slashes_key)
-        .await
-        .unwrap_or_default();
-    let bond_key = pos::bond_key(&BondId {
-        source: delegator.clone(),
-        validator: validator.clone(),
-    });
-    let epoched_bonds = query_storage_value::<C, Bonds>(client, &bond_key).await;
-    match epoched_bonds {
-        Some(epoched_bonds) => {
-            let mut delegated_amount: token::Amount = 0.into();
-            for bond in epoched_bonds.iter() {
-                let mut to_deduct = bond.neg_deltas;
-                for (epoch_start, &(mut delta)) in
-                    bond.pos_deltas.iter().sorted()
-                {
-                    // deduct bond's neg_deltas
-                    if to_deduct > delta {
-                        to_deduct -= delta;
-                        // If the whole bond was deducted, continue to
-                        // the next one
-                        continue;
-                    } else {
-                        delta -= to_deduct;
-                        to_deduct = token::Amount::default();
-                    }
-
-                    delta = apply_slashes(
-                        &slashes,
-                        delta,
-                        *epoch_start,
-                        None,
-                        None,
-                    );
-                    if epoch >= *epoch_start {
-                        delegated_amount += delta;
-                    }
-                }
-            }
-            Some(delegated_amount)
-        }
-        None => None,
-    }
+    namada::ledger::rpc::get_bond_amount_at(client, delegator, validator, epoch).await
 }
 
 pub async fn get_all_validators<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
     epoch: Epoch,
 ) -> HashSet<Address> {
-    unwrap_client_response::<C, _>(
-        RPC.vp()
-            .pos()
-            .validator_addresses(client, &Some(epoch))
-            .await,
-    )
+    namada::ledger::rpc::get_all_validators(client, epoch).await
 }
 
 pub async fn get_total_staked_tokens<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
     epoch: Epoch,
 ) -> token::Amount {
-    unwrap_client_response::<C, _>(
-        RPC.vp().pos().total_stake(client, &Some(epoch)).await,
-    )
+    namada::ledger::rpc::get_total_staked_tokens(client, epoch).await
 }
 
 async fn get_validator_stake<C: Client + namada::ledger::queries::Client + Sync>(
@@ -2647,61 +2290,18 @@ async fn get_validator_stake<C: Client + namada::ledger::queries::Client + Sync>
     epoch: Epoch,
     validator: &Address,
 ) -> token::Amount {
-    unwrap_client_response::<C, _>(
-        RPC.vp()
-            .pos()
-            .validator_stake(client, validator, &Some(epoch))
-            .await,
-    )
+    namada::ledger::rpc::get_validator_stake(client, epoch, validator).await
 }
 
 pub async fn get_delegators_delegation<C: Client + namada::ledger::queries::Client + Sync>(
     client: &C,
     address: &Address,
 ) -> HashSet<Address> {
-    unwrap_client_response::<C, _>(RPC.vp().pos().delegations(client, address).await)
+    namada::ledger::rpc::get_delegators_delegation(client, address).await
 }
 
 pub async fn get_governance_parameters<C: Client + namada::ledger::queries::Client + Sync>(client: &C) -> GovParams {
-    use namada::types::token::Amount;
-    let key = gov_storage::get_max_proposal_code_size_key();
-    let max_proposal_code_size = query_storage_value::<C, u64>(client, &key)
-        .await
-        .expect("Parameter should be definied.");
-
-    let key = gov_storage::get_max_proposal_content_key();
-    let max_proposal_content_size = query_storage_value::<C, u64>(client, &key)
-        .await
-        .expect("Parameter should be definied.");
-
-    let key = gov_storage::get_min_proposal_fund_key();
-    let min_proposal_fund = query_storage_value::<C, Amount>(client, &key)
-        .await
-        .expect("Parameter should be definied.");
-
-    let key = gov_storage::get_min_proposal_grace_epoch_key();
-    let min_proposal_grace_epochs = query_storage_value::<C, u64>(client, &key)
-        .await
-        .expect("Parameter should be definied.");
-
-    let key = gov_storage::get_min_proposal_period_key();
-    let min_proposal_period = query_storage_value::<C, u64>(client, &key)
-        .await
-        .expect("Parameter should be definied.");
-
-    let key = gov_storage::get_max_proposal_period_key();
-    let max_proposal_period = query_storage_value::<C, u64>(client, &key)
-        .await
-        .expect("Parameter should be definied.");
-
-    GovParams {
-        min_proposal_fund: u64::from(min_proposal_fund),
-        max_proposal_code_size,
-        min_proposal_period,
-        max_proposal_period,
-        max_proposal_content_size,
-        min_proposal_grace_epochs,
-    }
+    namada::ledger::rpc::get_governance_parameters(client).await
 }
 
 /// Try to find an alias for a given address from the wallet. If not found,
