+++ shared/src/ledger/mod.rs
@@ -13,6 +13,7 @@ pub mod storage;
 pub mod vp_host_fns;
 pub mod wallet;
 pub mod args;
+pub mod rpc;
 
 pub use namada_core::ledger::{
     gas, governance, parameters, storage_api, tx_env, vp_env,
